Configuraciones del la pagina de feed y register
