--- src/utils/utils.js
@@ -1,24 +1,0 @@
-export const bridgeData = (state) => {
-  const data = localStorage.getItem("data");
-  const dataJson = JSON.parse(data);
-  console.log(dataJson);
-  switch (state) {
-    case "ALLUSER":
-      setAllUser(dataJson);
-      localStorage.removeItem("data");
-      break;
-    default:
-      break;
-  }
-};
-
-export const login = (data) => {
-  localStorage.setItem("user", data);
-  const parseUser = JSON.parse(data);
-  setUser(parseUser);
-};
-
-export const logout = () => {
-  localStorage.removeItem("user");
-  setUser(null);
-};
